requirements.txt
biopython
progressbar2